app/pipeline/inference/modules/awsLocalizationModule/requirements.txt
boto3==1.26.118
numpy==1.22.4
opencv-python-headless==4.8.1.78
py360convert==0.1.0
pika==1.3.2
python-decouple==3.8
requests==2.31.0